Major dependency/module reformatting
